Move LayoutMiscTab to its own .cpp and .h files
