+++ foo_ui_columns/config_layout.cpp
@@ -1,8 +1,8 @@
 #include "stdafx.h"
 #include "layout.h"
 #include "config.h"
-#include "splitter.h"
 #include "splitter_utils.h"
+#include "tab_layout_misc.h"
 
 namespace cui::prefs {
 constexpr GUID layout_page_guid = {0x1d904b23, 0x9357, 0x4cf7, {0xbf, 0xdf, 0xd3, 0x9e, 0xa0, 0x30, 0x73, 0xb1}};
@@ -1091,94 +1091,7 @@ private:
     cui::prefs::PreferencesTabHelper m_helper{IDC_TITLE1};
 };
 
-class LayoutMiscTab : public preferences_tab {
-public:
-    HWND create(HWND wnd) override
-    {
-        return m_helper.create(
-            wnd, IDD_LAYOUT_MISC, [this](auto&&... args) { return ConfigProc(std::forward<decltype(args)>(args)...); });
-    }
-
-    const char* get_name() override { return "Misc"; }
-
-    bool get_help_url(pfc::string_base& p_out) override
-    {
-        p_out = "http://yuo.be/wiki/columns_ui:config:layout";
-        return true;
-    }
-
-private:
-    BOOL ConfigProc(HWND wnd, UINT msg, WPARAM wp, LPARAM lp)
-    {
-        switch (msg) {
-        case WM_INITDIALOG: {
-            SetDlgItemInt(wnd, IDC_SHOW_DELAY, cfg_sidebar_show_delay, FALSE);
-            SetDlgItemInt(wnd, IDC_HIDE_DELAY, cfg_sidebar_hide_delay, FALSE);
-            EnableWindow(GetDlgItem(wnd, IDC_SHOW_DELAY_SPIN), cfg_sidebar_use_custom_show_delay);
-            EnableWindow(GetDlgItem(wnd, IDC_SHOW_DELAY), cfg_sidebar_use_custom_show_delay);
-            SendDlgItemMessage(wnd, IDC_USE_CUSTOM_SHOW_DELAY, BM_SETCHECK, cfg_sidebar_use_custom_show_delay, 0);
-            SendDlgItemMessage(
-                wnd, IDC_ALLOW_LOCKED_PANEL_RESIZING, BM_SETCHECK, settings::allow_locked_panel_resizing, 0);
-
-            SendDlgItemMessage(wnd, IDC_SHOW_DELAY_SPIN, UDM_SETRANGE32, 0, 10000);
-            SendDlgItemMessage(wnd, IDC_HIDE_DELAY_SPIN, UDM_SETRANGE32, 0, 10000);
-
-            HWND wnd_custom_divider_width_spin = GetDlgItem(wnd, IDC_CUSTOM_DIVIDER_WIDTH_SPIN);
-
-            SetDlgItemInt(wnd, IDC_CUSTOM_DIVIDER_WIDTH, settings::custom_splitter_divider_width, FALSE);
-            SendMessage(wnd_custom_divider_width_spin, UDM_SETRANGE32, 0, 20);
-
-            m_initialised = true;
-        } break;
-        case WM_DESTROY:
-            m_initialised = false;
-            break;
-        case WM_COMMAND:
-            switch (wp) {
-            case (EN_CHANGE << 16) | IDC_HIDE_DELAY:
-                if (m_initialised) {
-                    BOOL result;
-                    int new_height = GetDlgItemInt(wnd, LOWORD(wp), &result, FALSE);
-                    if (result)
-                        cfg_sidebar_hide_delay = new_height;
-                }
-                break;
-            case (EN_CHANGE << 16) | IDC_SHOW_DELAY:
-                if (m_initialised) {
-                    BOOL result;
-                    int new_height = GetDlgItemInt(wnd, LOWORD(wp), &result, FALSE);
-                    if (result)
-                        cfg_sidebar_show_delay = new_height;
-                }
-                break;
-            case (EN_CHANGE << 16) | IDC_CUSTOM_DIVIDER_WIDTH:
-                if (m_initialised) {
-                    BOOL result;
-                    int new_width = GetDlgItemInt(wnd, LOWORD(wp), &result, FALSE);
-                    if (result)
-                        settings::custom_splitter_divider_width = new_width;
-                    splitter_window_impl::g_on_size_change();
-                }
-                break;
-            case IDC_USE_CUSTOM_SHOW_DELAY:
-                cfg_sidebar_use_custom_show_delay = SendMessage((HWND)lp, BM_GETCHECK, 0, 0);
-                EnableWindow(GetDlgItem(wnd, IDC_SHOW_DELAY_SPIN), cfg_sidebar_use_custom_show_delay);
-                EnableWindow(GetDlgItem(wnd, IDC_SHOW_DELAY), cfg_sidebar_use_custom_show_delay);
-                break;
-            case IDC_ALLOW_LOCKED_PANEL_RESIZING:
-                settings::allow_locked_panel_resizing = Button_GetCheck((HWND)lp) == BST_CHECKED;
-                break;
-            }
-            break;
-        }
-        return 0;
-    }
-
-    bool m_initialised{};
-    cui::prefs::PreferencesTabHelper m_helper{{IDC_TITLE1, IDC_TITLE2}};
-};
-
-LayoutMiscTab g_tab_layout_misc;
+cui::prefs::LayoutMiscTab g_tab_layout_misc;
 LayoutTab g_tab_layout;
 
 preferences_tab* tabs_layout[] = {&g_tab_layout, &g_tab_layout_misc};
